+++ src/databricks/azext_databricks/__init__.py
@@ -12,10 +12,8 @@ class DatabricksClientCommandsLoader(AzCommandsLoader):
 
     def __init__(self, cli_ctx=None):
         from azure.cli.core.commands import CliCommandType
-        from azext_databricks._client_factory import cf_databricks
         databricks_custom = CliCommandType(
-            operations_tmpl='azext_databricks.custom#{}',
-            client_factory=cf_databricks)
+            operations_tmpl='azext_databricks.custom#{}')
         super().__init__(cli_ctx=cli_ctx,
                          custom_command_type=databricks_custom)